Calculator Example (#284)
In this example, we illustrate utilizing the calculator to compute a derived constant value (Average Pressure) on a derived part (a clip plane). When then also illustrate creating a query of that constant over time, to understand its variation with time. The key aspect shown here is how to perform calculator operation, along with illustrating that how that newly created constant varies over time. We are extracting the useful data from our simulation to better understand its performance.